+++ app/src/main/java/campus/tech/kakao/map/data/di/NetworkModule.kt
@@ -1,0 +1,32 @@
+package campus.tech.kakao.map.data.di
+
+import campus.tech.kakao.map.data.remote.api.KakaoApiService
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NetworkModule {
+
+    private const val BASE_URL = "https://dapi.kakao.com/"
+
+    @Provides
+    @Singleton
+    fun provideRetrofit(): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideKakaoApiService(retrofit: Retrofit): KakaoApiService {
+        return retrofit.create(KakaoApiService::class.java)
+    }
+}
