--- app/src/main/java/campus/tech/kakao/map/data/remote/api/RetrofitClient.kt
@@ -1,15 +1,0 @@
-package campus.tech.kakao.map.data.remote.api
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitClient {
-    private const val BASE_URL = "https://dapi.kakao.com/"
-
-    val retrofitService: KakaoApiService by lazy { Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-        .create(KakaoApiService::class.java)
-    }
-}
